# plugins/dev-lifecycle/templates/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI block",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users/{user_id}/ban": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Ban user (admin)",
        "description": "Valid from `status in {\"active\", \"suspended\"}` -- an already-`banned`\nuser raises `ConflictError` (409, idempotent re-ban is rejected rather\nthan silently no-op'd, matching `suspend`'s own strict-transition\nposture). Self-protection: the acting admin cannot ban themselves (409).\nSame refresh-token revocation as `suspend` above -- see that handler's\nown docstring.",
        "operationId": "ban_admin_user_admin_users__user_id__ban_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The action conflicts with the user's current state, or is a self-protection guard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/flags": {
      "get": {
        "tags": [
          "moderation"
        ],
        "summary": "List flags (admin)",
        "description": "`?status=`/`?target_type=` each filter to one exact value, composable\n\u2014 the queue's primary \"show me open reports\" filter shape.",
        "operationId": "list_admin_flags_admin_flags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FlagStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FlagTargetType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FlagOut_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/flags/{flag_id}": {
      "get": {
        "tags": [
          "moderation"
        ],
        "summary": "Get flag (admin)",
        "operationId": "get_admin_flag_admin_flags__flag_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Flag (or its target/author) not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/flags/{flag_id}/resolve": {
      "post": {
        "tags": [
          "moderation"
        ],
        "summary": "Resolve flag (admin)",
        "description": "State machine: only an `open` flag can be resolved (409 otherwise).\nDispatches on `payload.action` \u2014 see this module's own docstring for\nthe full per-action, per-`target_type` behavior. The content/author\nside effect (if any) runs BEFORE the flag itself is marked resolved,\nso a 404/409/422 raised by that side effect leaves the flag untouched\n(still `open`), never resolved-with-a-failed-action.",
        "operationId": "resolve_admin_flag_admin_flags__flag_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagResolveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Flag (or its target/author) not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The flag isn't 'open', or the action is a self-protection guard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/flags/{flag_id}/dismiss": {
      "post": {
        "tags": [
          "moderation"
        ],
        "summary": "Dismiss flag (admin)",
        "description": "State machine: only an `open` flag can be dismissed (409 otherwise).\nNever performs a content/author action \u2014 `dismiss` is \"this report\ndoesn't warrant action,\" not a moderation outcome.",
        "operationId": "dismiss_admin_flag_admin_flags__flag_id__dismiss_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagDismissIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the caller lacks the 'admin' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Flag (or its target/author) not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The flag isn't 'open', or the action is a self-protection guard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminUserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email",
          "roles",
          "status",
          "email_verified",
          "created_at"
        ],
        "title": "AdminUserOut",
        "description": "The read shape every admin user-management endpoint returns.\nDeliberately NO `password_hash`, NO token fields \u2014 this is the ONE\nplace `app/models/user.py`'s `User` ORM instance is ever serialized for\nan admin caller, and leaking either would be a straightforward secret\nleak into an HTTP response (and, via `app/api/routers/admin.py`'s own\n`audit_event(...)` calls, never into an audit log either \u2014 see that\nrouter's module docstring). `from_attributes=True` lets\n`AdminUserOut.model_validate(user)` read straight off the ORM\ninstance's attributes, the same convention `app/schemas/item.py`'s\n`ItemOut` documents."
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "THE error envelope every error response in this app uses \u2014 every\nnon-2xx error, including request-boundary validation failures once\nStep 2's `RequestValidationError` handler remaps them (see this\nmodule's docstring):\n\n    {\"error\": {\"code\": \"not_found\", \"message\": \"...\", \"details\": null}}\n\nEvery field required except `details`, which is `null` (omitted from\nthe exception's own `to_envelope()` output) when there's nothing more\nspecific than the top-level `code`/`message` to say.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FlagOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "target_type": {
            "$ref": "#/components/schemas/FlagTargetType"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "reporter_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/FlagStatus"
          },
          "resolved_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By Id"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "target_type",
          "target_id",
          "reporter_id",
          "reason",
          "status",
          "resolved_by_id",
          "resolved_at",
          "resolution_note",
          "created_at"
        ],
        "title": "FlagOut",
        "description": "The read shape every moderation admin endpoint returns.\n`from_attributes=True` lets `FlagOut.model_validate(flag)` read\nstraight off the ORM instance, the same convention `app/schemas/\nadmin.py`'s `AdminUserOut`/`app/schemas/blog.py`'s `BlogPostSummaryOut`\ndocument."
      },
      "FlagResolveIn": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ResolveAction"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "FlagResolveIn",
        "description": "`POST /admin/flags/{flag_id}/resolve`'s request body \u2014 `action`\npicks which content/author-side effect (if any) the resolve also\nperforms (see `ResolveAction`'s own docstring); `note` is an optional\nfree-text resolution note, persisted verbatim to `Flag.\nresolution_note` (never rendered as markup anywhere in this app \u2014 same\n\"opaque, never sanitized\" posture `app/models/comment.py`'s own\ndocstring documents for a plain-text column with no render path)."
      },
      "FlagStatus": {
        "type": "string",
        "enum": [
          "open",
          "resolved",
          "dismissed"
        ],
        "title": "FlagStatus"
      },
      "FlagTargetType": {
        "type": "string",
        "enum": [
          "blog_post",
          "comment",
          "user"
        ],
        "title": "FlagTargetType"
      },
...
      "Page_FlagOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FlagOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[FlagOut]"
      },
...
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "banned"
        ],
        "title": "UserStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}