# plugins/dev-lifecycle/templates/packages/api-client/src/generated/models/flagStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * FastAPI block
 * OpenAPI spec version: 0.1.0
 */

export type FlagStatus = typeof FlagStatus[keyof typeof FlagStatus];


export const FlagStatus = {
  open: 'open',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;